pyproject.toml
[tool.poetry]
name = "paygate-api-client"
version = "v1"
description = "A client library for accessing Paygate API"

authors = []

readme = "README.md"
packages = [
    {include = "paygate_api_client"},
]
include = ["CHANGELOG.md", "paygate_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.6"
httpx = "^0.15.0"
attrs = "^20.1.0"
python-dateutil = "^2.8.1"

[tool.black]
line-length = 120
target_version = ['py36', 'py37', 'py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"